staged

+++ homeworks/project/src/hw6/Components/DenoiseData.h
@@ -26,7 +26,7 @@ struct DenoiseData {
 	Ubpa::Utopia::Mesh copy;
 
 	[[UInspector::tooltip("number of iterations")]]
-	float k;
+	int k;
 	
 	[[UInspector::tooltip("step")]]
 	float lambda{ 0.1f };
